MAHOUT-405 - Undeprecate a number of interfaces and all of the random.engine classes (some by deletion).
This involved simply undeprecating most of the members of math.function by inspection for math.function.Functions or because the rest are interfaces anyway.

For MersenneTwister, I went back to the C reference implementation and generated unit tests based on that.  Then I walked the code generating unit tests based on equivalent performance.


git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@950378 13f79535-47bb-0310-9956-ffa450edef68