CLOUDERA-BUILD. CDH-specific build tweaks.
Don't bother with the .zip and .tar.bz2 mahout-distribution assemblies
- we don't need them for the build, and they just take
time/space/etc. Removing them decreases the size of our source
packages and binary tarball considerably.

Also set the hadoop.version property by default to the CDH version of
Hadoop, so that it doesn't need to be explicitly specified in the mvn call.