MAHOUT-933: - refactored ClusteringPolicies into hierarchy under new AbstractClusteringPolicy - added close() to ClusteringPolicy to allow policy-specific actions needed to compute convergence - removed ClusteringPolicy from ClusterIterator constructor as ClusterClassifier already has one - added convergence computations for kmeans and fuzzyk - added final clustersOut renaming to add -final suffix - updated Display examples and unit tests to reflect above - all tests run
I think it is time to begin refactoring the buildClusters methods of the respective clustering drivers to use ClusterIterator as it seems to be producing equivalent results to the original implementations. This will involve removing a lot of existing driver, mapper and reducer code and many time-consuming unit tests. It will also have some impact on other components as the representation of clusters in the file system changes from Cluster to self-describing ClusterWritable.

git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@1298625 13f79535-47bb-0310-9956-ffa450edef68