MAHOUT-1223: Fixed point being skipped in StreamingKMeans when iterating through centroids from a reducer
When calling StreamingKMeans in the reducer (to collapse the number of clusters
to they can fit into memory), the clustering is done on the Hadoop reducer
iterable.
Currently, the first Centroid is added directly as a special case and then is
skipped when iterating through the main loop.
However, Hadoop reducer iterables cannot be rewound therefore causing SKM to
skip one point.



git-svn-id: https://svn.apache.org/repos/asf/mahout/trunk@1484747 13f79535-47bb-0310-9956-ffa450edef68